refactor: organize components about `popup`
