+++ apps/calendar/src/components/floatingLayer.tsx
@@ -1,8 +1,8 @@
 import { FunctionComponent, h } from 'preact';
 
-import DetailPopup from '@src/components/popup/detailPopup';
+import { EventDetailPopup } from '@src/components/popup/eventDetailPopup';
 import { EventFormPopup } from '@src/components/popup/eventFormPopup';
-import SeeMorePopup from '@src/components/popup/seeMorePopup';
+import { SeeMoreEventsPopup } from '@src/components/popup/seeMoreEventsPopup';
 import { useStore } from '@src/contexts/calendarStore';
 import { topLevelStateSelector } from '@src/selectors';
 import { PopupType } from '@src/slices/popup';
@@ -13,11 +13,11 @@ import { DetailPopupParam, EventFormPopupParam, PopupParamMap, SeeMorePopupParam
 const renderPopup = (popupType: PopupType | null, param: PopupParamMap[PopupType]) => {
   switch (popupType) {
     case PopupType.seeMore:
-      return <SeeMorePopup {...(param as SeeMorePopupParam)} />;
+      return <SeeMoreEventsPopup {...(param as SeeMorePopupParam)} />;
     case PopupType.form:
       return <EventFormPopup {...(param as EventFormPopupParam)} />;
     case PopupType.detail:
-      return <DetailPopup {...(param as DetailPopupParam)} />;
+      return <EventDetailPopup {...(param as DetailPopupParam)} />;
     default:
       return null;
   }
