+++ apps/calendar/src/components/popup/eventDetailPopup.tsx
@@ -2,9 +2,7 @@ import { FunctionComponent, h } from 'preact';
 
 import { DetailPopupParam } from '@t/store';
 
-const DetailPopup: FunctionComponent<DetailPopupParam> = () => {
+export const EventDetailPopup: FunctionComponent<DetailPopupParam> = () => {
   // @TODO: detail popup
   return <div>Detail Popup</div>;
 };
-
-export default DetailPopup;
